Delete Method Updated

+++ uit_fastapi/fastapi.py
@@ -123,7 +123,9 @@ def delete_hero_by_id(todo_id: int, session: session):
 @app.delete('/todo')
 def delete_all(session: session):
     all_content = session.exec(select(Todo)).all()
-    for content in all_content:
-        session.delete(content)
-    session.commit()
-    return {"status": "All content deleted"}
+    if all_content:
+        for content in all_content:
+            session.delete(content)
+        session.commit()
+        return {"status": "All content deleted"}
+    return {"message":"No content Found"}
